# topologically ordered update cycle