--- a/build_npm.ts
+++ b/build_npm.ts
@@ -88,10 +88,10 @@ await build({
 	mappings: Object.fromEntries(
 		["binder", "builtin_aliases_deps", "lambda", "struct", "typedefs",].map((submodule_path) => {
 			return [
-				"https://deno.land/x/kitchensink_ts@v0.7.2/" + submodule_path + ".ts",
+				"https://deno.land/x/kitchensink_ts@v0.7.3/" + submodule_path + ".ts",
 				{
 					name: "kitchensink_ts",
-					version: "^v0.7.2",
+					version: "^v0.7.3",
 					subPath: submodule_path,
 				}
 			]