--- a/deno.json
+++ b/deno.json
@@ -1,6 +1,6 @@
 {
 	"name": "tsignal_ts",
-	"version": "0.1.3",
+	"version": "0.2.0",
 	"description": "a topological order respecting signals library inspired by SolidJS",
 	"author": "Omar Azmi",
 	"license": "Lulz plz don't steal yet",
@@ -21,7 +21,7 @@
 		"allowJs": true
 	},
 	"dependencies": {
-		"kitchensink_ts": "^0.7.2"
+		"kitchensink_ts": "^0.7.3"
 	},
 	"devDependencies": {
 		"typescript": "^5.0.0",