--- a/src/context.ts
+++ b/src/context.ts
@@ -3,7 +3,7 @@
 */
 
 import { DEBUG, bindMethodToSelfByName, bind_array_clear, bind_array_pop, bind_array_push, bind_map_clear, bind_map_delete, bind_map_get, bind_map_set, bind_set_add, bind_set_clear, bind_set_delete, bind_set_has } from "./deps.ts"
-import { hash_ids, symmetric_difference_of_sets } from "./funcdefs.ts"
+import { hash_ids } from "./funcdefs.ts"
 import { EffectFn, MemoFn, SimpleSignalInstance } from "./signal.ts"
 import { EqualityFn, FROM_ID, HASHED_IDS, ID, Signal, SignalClass, SignalUpdateStatus, TO_ID, UNTRACKED_ID } from "./typedefs.ts"
 
@@ -53,126 +53,177 @@ export class Context {
 			rmap_delete = bind_map_delete(rmap)
 
 		const
-			ids_to_visit_cache = new Map<HASHED_IDS, Set<ID>>(),
+			ids_to_visit_cache = new Map<HASHED_IDS, ID[]>(),
 			ids_to_visit_cache_get = bind_map_get(ids_to_visit_cache),
 			ids_to_visit_cache_set = bind_map_set(ids_to_visit_cache),
-			ids_to_visit_cache_clear = bind_map_clear(ids_to_visit_cache),
-			ids_to_visit_cache_create_new_entry = (source_ids: ID[]): Set<ID> => {
-				const
-					to_visit = new Set<ID>(),
-					to_visit_add = bind_set_add(to_visit),
-					to_visit_has = bind_set_has(to_visit)
-				const dfs_visiter = (id: ID) => {
-					if (!to_visit_has(id)) {
-						to_visit_add(id)
-						fmap_get(id)?.forEach(dfs_visiter)
-					}
+			ids_to_visit_cache_clear = bind_map_clear(ids_to_visit_cache)
+
+		/** creates a topologically ordered array of ids to visit, when propagation is initiated from `source_ids`. <br>
+		 * the function ensures that `source_ids` are always put at the very beginning of the retuned value,
+		 * as this kind of ordering is of utter importance to the {@link fireUpdateCycle | `fireUpdateCycle`} function, which uses this.
+		 * @example
+		 * ```ts
+		 * // assume our directed acyclic graph is:
+		 * // A->[B,C] ; B->[D, F] ; C->[E, F] ; D->[F, G] ; E->[F, H] ; F->[I] ; G->[I] ; H->[I] ; I->[J] ;
+		 * console.log(ids_to_visit_cache_create_new_entry([A])) // [A, C, E, H, B, D, F, G, I, J]
+		 * console.log(ids_to_visit_cache_create_new_entry([B, E])) // [E, B, D, F, G, H, I, J]
+		 * // notice that B and E were placed first, even though enforcing it was not necessary for the array to be still considered as topologically ordered.
+		 * ```
+		*/
+		const ids_to_visit_cache_create_new_entry = (source_ids: ID[]): ID[] => {
+			const
+				to_visit = new Set<ID>(),
+				to_visit_add = bind_set_add(to_visit),
+				to_visit_has = bind_set_has(to_visit)
+			const dfs_visitor = (id: ID) => {
+				if (!to_visit_has(id)) {
+					fmap_get(id)?.forEach(dfs_visitor)
+					to_visit_add(id)
 				}
-				source_ids.forEach(dfs_visiter)
-				return to_visit
-			},
-			get_ids_to_visit = (...source_ids: ID[]): Set<ID> => {
-				const hash = hash_ids(source_ids)
-				return ids_to_visit_cache_get(hash) ?? (
-					ids_to_visit_cache_set(hash, ids_to_visit_cache_create_new_entry(source_ids)) &&
-					ids_to_visit_cache_get(hash)!
-				)
 			}
+			source_ids.forEach(dfs_visitor)
+			// delete the `source_ids` from `to_visit`, as will add them later to the very beginning of the return array
+			source_ids.forEach(bind_set_delete(to_visit))
+			// currently, `to_visit` is in reverse-topological order, which is what you get with the type of DFS we're doing here.
+			// so reverse `to_visit` to get a topologically ordered set
+			return [...to_visit, ...source_ids].reverse()
+		}
+
+		/** get a topologically ordered set of ids to visit, when propagation is initiated from `source_ids`.
+		 * the result is cached for quicker followup retrievals.
+		 * see {@link ids_to_visit_cache_create_new_entry | `ids_to_visit_cache_create_new_entry`} for a better explanation.
+		*/
+		const get_ids_to_visit = (...source_ids: ID[]): ID[] => {
+			const hash = hash_ids(source_ids)
+			return ids_to_visit_cache_get(hash) ?? (
+				ids_to_visit_cache_set(hash, ids_to_visit_cache_create_new_entry(source_ids)) &&
+				ids_to_visit_cache_get(hash)!
+			)
+		}
 
 		const
+			/** contains a mapping of all signal `ID`s and their {@link Signal | signal instances} */
 			all_signals = new Map<ID, Signal<any>>(),
 			all_signals_get = bind_map_get(all_signals),
 			all_signals_set = bind_map_set(all_signals),
 			all_signals_delete = bind_map_delete(all_signals)
 
 		const
-			to_visit_this_cycle = new Set<ID>(),
-			to_visit_this_cycle_add = bind_set_add(to_visit_this_cycle),
-			to_visit_this_cycle_delete = bind_set_delete(to_visit_this_cycle),
-			to_visit_this_cycle_clear = bind_set_clear(to_visit_this_cycle),
-			updated_this_cycle = new Map<ID, SignalUpdateStatus>(),
-			updated_this_cycle_get = bind_map_get(updated_this_cycle),
-			updated_this_cycle_set = bind_map_set(updated_this_cycle),
-			updated_this_cycle_clear = bind_map_clear(updated_this_cycle),
+			/** this `Set` of `ID`s is essentially a dynamic book-keeper for which signal `ID`s have had their dependency-signals declare an update,
+			 * and as a result, these set of `ID`s are now awaiting in queue to be executed themselves, in the follow-up update loop.
+			 * this technique works correctly, particularly because {@link fireUpdateCycle |update cycle's} loop is topologically ordered.
+			 * and so, we are guaranteed to have encountered all of a certain `ID`'s dependency-signals, before we test for its existence **here** via deletion.
+			*/
+			next_to_visit_this_cycle = new Set<ID>(),
+			next_to_visit_this_cycle_add = bind_set_add(next_to_visit_this_cycle),
+			next_to_visit_this_cycle_delete = bind_set_delete(next_to_visit_this_cycle),
+			next_to_visit_this_cycle_clear = bind_set_clear(next_to_visit_this_cycle)
+
+		const
+			// TODO: consider whether or not aborted signals should poison only their immediate observers, or have it propagate deeply into all consequent observers.
+			/** this is very similar to {@link next_to_visit_this_cycle}, except, it book-keeps which signals have been declared aborted.
+			 * aborted signals are never run, and they poison their immediate observers to also become aborted and not run (even if they have a dependency that has been updated).
+			*/
+			not_to_visit_this_cycle = new Set<ID>(),
+			not_to_visit_this_cycle_add = bind_set_add(not_to_visit_this_cycle),
+			not_to_visit_this_cycle_has = bind_set_has(not_to_visit_this_cycle),
+			not_to_visit_this_cycle_clear = bind_set_clear(not_to_visit_this_cycle)
+
+		const
+			/** when {@link DEBUG.LOG | debug-logging} is turned on, this gives a summary of what has been updated,
+			 * and what topologically-ordered trajectory was taken for the update.
+			*/
+			status_this_cycle = /* @__PURE__ */ new Map<ID, SignalUpdateStatus>(),
+			status_this_cycle_set = /* @__PURE__ */ bind_map_set(status_this_cycle),
+			status_this_cycle_clear = /* @__PURE__ */ bind_map_clear(status_this_cycle)
+
+		const
+			/** certain signals have cleaning up to do *after* each update cycle.
+			 * those callback functions are stored here, and then called in reverse order
+			*/
 			postruns_this_cycle: ID[] = [],
 			postruns_this_cycle_push = bind_array_push(postruns_this_cycle),
-			postruns_this_cycle_clear = bind_array_pop(postruns_this_cycle),
+			postruns_this_cycle_clear = bind_array_pop(postruns_this_cycle)
+
+		const fireUpdateCycle = (...source_ids: FROM_ID[]) => {
+			// clear up all book-keeping artifacts from the any previous update cycle, and start anew.
+			next_to_visit_this_cycle_clear()
+			not_to_visit_this_cycle_clear()
+			if (DEBUG.LOG) { /* @__PURE__ */ status_this_cycle_clear() }
+			source_ids.forEach(next_to_visit_this_cycle_add)
+			/** some signals make use of the information whether of not they are the source of a cycle via the optional `forced` argument of {@link Signal.run | `Signal.run`}.
+			 * so we simply take a note of how many source signals we begun with, and decrement it each time in the loop.
+			 * this works because the first few items of {@link topological_ids} is guaranteed to be the {@link source_ids},
+			 * thanks to {@link get_ids_to_visit} (which guarantees this behavior).
+			 * the `forced` parameter is of particular importance to signals that can fire on their own, such as state signal and effect signals in [`./signal.ts`](./signal.ts).
+			*/
+			let number_of_forced_ids = source_ids.length
+			const topological_ids = get_ids_to_visit(...source_ids)
+			for (const source_id of topological_ids) {
+				if (
+					next_to_visit_this_cycle_delete(source_id) &&
+					!not_to_visit_this_cycle_has(source_id)
+				) {
+					const signal_update_status = executeSignal(source_id, number_of_forced_ids-- > 0)
+					if (signal_update_status !== SignalUpdateStatus.UNCHANGED) {
+						fmap_get(source_id)?.forEach(
+							signal_update_status >= SignalUpdateStatus.UPDATED ?
+								next_to_visit_this_cycle_add :
+								not_to_visit_this_cycle_add
+						)
+					}
+					if (DEBUG.LOG) { status_this_cycle_set(source_id, signal_update_status) }
+				}
+				// if there are no remaining ids `to_visit_this_cycle`, then it is our cue to terminate early,
+				// since the remaining `source_id`s will not be part of the update.
+				if (next_to_visit_this_cycle.size <= 0) { break }
+			}
+			if (DEBUG.LOG) {
+				console.log("topological visiting ordering: ", [...status_this_cycle].map(([id, status]) => {
+					return [all_signals_get(id)!.name, status]
+				}))
+			}
+			// run all of the `postrun` cleanup methods after the end of this cycle (in reverse order of accumulation)
+			if (DEBUG.LOG) { console.log("UPDATE_POSTRUNS:\t", postruns_this_cycle) }
+			let postrun_id: ID | undefined
+			while (postrun_id = postruns_this_cycle_clear()) {
+				all_signals_get(postrun_id)?.postrun!()
+			}
+		}
+
+		const executeSignal = (id: ID, force?: true | any): SignalUpdateStatus => {
+			const
+				forced = force === true,
+				this_signal = all_signals_get(id),
+				this_signal_update_status = this_signal?.run(forced) ?? SignalUpdateStatus.UNCHANGED
+			if (
+				this_signal_update_status >= SignalUpdateStatus.UPDATED &&
+				this_signal!.postrun
+			) { postruns_this_cycle_push(id) }
+			return this_signal_update_status
+		}
+
+		const
+			/** this is used by the {@link batch | batching-functions} to collect source `ID`s, and then dispatch all at once. */
 			batched_ids: FROM_ID[] = [],
 			batched_ids_push = bind_array_push(batched_ids),
-			batched_ids_clear = bind_array_clear(batched_ids),
-			fireUpdateCycle = (...source_ids: FROM_ID[]) => {
-				to_visit_this_cycle_clear()
-				updated_this_cycle_clear()
-				// clone the ids to visit into the "visit cycle for this update"
-				get_ids_to_visit(...source_ids).forEach(to_visit_this_cycle_add)
-				// fire the signal and propagate its reactivity.
-				// the souce signals are `force`d in order to skip any unresolved dependency check.
-				// this is needed because although state signals do not have dependencies, effect signals may have one.
-				// but effect signals are themselves designed to be fired/ran as standalone signals
-				for (const source_id of source_ids) {
-					propagateSignalUpdate(source_id, true)
-				}
-				// run all of the `postrun` cleanup methods after the end of this cycle (in reverse order of accumulation)
-				if (DEBUG.LOG) { console.log("UPDATE_POSTRUNS:\t", postruns_this_cycle) }
-				let postrun_id: ID | undefined
-				while (postrun_id = postruns_this_cycle_clear()) {
-					all_signals_get(postrun_id)?.postrun!()
-				}
-			},
-			startBatching = () => (++batch_nestedness),
-			endBatching = () => {
-				if (--batch_nestedness <= 0) {
-					batch_nestedness = 0
-					fireUpdateCycle(...batched_ids_clear())
-				}
-			},
-			scopedBatching = <T extends any = void, ARGS extends any[] = []>(
-				fn: (...args: ARGS) => T,
-				...args: ARGS
-			): T => {
-				startBatching()
-				const return_value = fn(...args)
-				endBatching()
-				return return_value
-			}
+			batched_ids_clear = bind_array_clear(batched_ids)
 
-		const propagateSignalUpdate = (id: ID, force?: true | any) => {
-			if (to_visit_this_cycle_delete(id)) {
-				const
-					forced = force === true,
-					this_signal = all_signals_get(id)
-				if (DEBUG.LOG) { console.log("UPDATE_CYCLE\t", "visiting   :\t", this_signal?.name) }
-				// first make sure that all of this signal's dependencies are up to date (should they be among the set of ids to visit this update cycle)
-				// `any_updated_dependency` is always initially `false`. however, if the signal id was `force`d, then it would be `true`, and skip dependency checking and updating.
-				// you must use `force === true` for `StateSignal`s (because they are dependency free), or for a independently fired `EffectSignal`
-				let any_updated_dependency: SignalUpdateStatus = forced ? SignalUpdateStatus.UPDATED : SignalUpdateStatus.UNCHANGED
-				if (any_updated_dependency <= SignalUpdateStatus.UNCHANGED) {
-					for (const dependency_id of rmap_get(id) ?? []) {
-						propagateSignalUpdate(dependency_id)
-						any_updated_dependency |= updated_this_cycle_get(dependency_id) ?? SignalUpdateStatus.UNCHANGED
-					}
-				}
-				// now, depending on two AND criterias:
-				// 1) at least one dependency has updated (or must be free of dependencies via `force === true`)
-				// 2) AND, this signal's value has changed after the update computation (ie `run()` method)
-				// if both criterias are met, then this signal should propagate forward towards its observers
-				let this_signal_update_status: SignalUpdateStatus = any_updated_dependency
-				if (this_signal_update_status >= SignalUpdateStatus.UPDATED) {
-					this_signal_update_status = this_signal?.run(forced) ?? SignalUpdateStatus.UNCHANGED
-					/* I think we should ignore batching aborted source-signals here. instead, we should let the source itself handle how it wishes to be batched, or when it wishes to run
-					if (this_signal_update_status <= SignalUpdateStatus.ABORTED) {
-						// we only batch source ids which result in an ABORTED signal, rather than the collateral resulting ABORTED signals as a consequence
-						batched_ids_push(id)
-					}
-					*/
-				}
-				updated_this_cycle_set(id, this_signal_update_status)
-				if (DEBUG.LOG) { console.log("UPDATE_CYCLE\t", this_signal_update_status > 0 ? "propagating:\t" : this_signal_update_status < 0 ? "delaying    \t" : "blocking   :\t", this_signal?.name) }
-				if (this_signal_update_status >= SignalUpdateStatus.UPDATED) {
-					if (this_signal!.postrun) { postruns_this_cycle_push(id) }
-					fmap_get(id)?.forEach(propagateSignalUpdate)
-				}
+		const startBatching = () => (++batch_nestedness)
+		const endBatching = () => {
+			if (--batch_nestedness <= 0) {
+				batch_nestedness = 0
+				fireUpdateCycle(...batched_ids_clear())
 			}
+		}
+		const scopedBatching = <T extends any = void, ARGS extends any[] = []>(
+			fn: (...args: ARGS) => T,
+			...args: ARGS
+		): T => {
+			startBatching()
+			const return_value = fn(...args)
+			endBatching()
+			return return_value
 		}
 
 		this.addEdge = (src_id: FROM_ID, dst_id: TO_ID): boolean => {