--- a/src/deps.ts
+++ b/src/deps.ts
@@ -11,11 +11,11 @@ export {
 	bind_set_clear,
 	bind_set_delete,
 	bind_set_has
-} from "https://deno.land/x/kitchensink_ts@v0.7.2/binder.ts"
-export { array_isArray, noop, object_assign, object_keys, object_values, promise_forever, promise_reject, promise_resolve } from "https://deno.land/x/kitchensink_ts@v0.7.2/builtin_aliases_deps.ts"
-export { THROTTLE_REJECT, throttle, throttleAndTrail } from "https://deno.land/x/kitchensink_ts@v0.7.2/lambda.ts"
-export { prototypeOfClass } from "https://deno.land/x/kitchensink_ts@v0.7.2/struct.ts"
-export type { CallableFunctionsOf, ConstructorOf, MethodsOf, StaticImplements } from "https://deno.land/x/kitchensink_ts@v0.7.2/typedefs.ts"
+} from "https://deno.land/x/kitchensink_ts@v0.7.3/binder.ts"
+export { array_isArray, noop, object_assign, object_keys, object_values, promise_forever, promise_reject, promise_resolve } from "https://deno.land/x/kitchensink_ts@v0.7.3/builtin_aliases_deps.ts"
+export { THROTTLE_REJECT, throttle, throttleAndTrail } from "https://deno.land/x/kitchensink_ts@v0.7.3/lambda.ts"
+export { prototypeOfClass } from "https://deno.land/x/kitchensink_ts@v0.7.3/struct.ts"
+export type { CallableFunctionsOf, ConstructorOf, MethodsOf, StaticImplements } from "https://deno.land/x/kitchensink_ts@v0.7.3/typedefs.ts"
 
 export const enum DEBUG {
 	LOG = 0,