--- a/src/funcdefs.ts
+++ b/src/funcdefs.ts
@@ -10,12 +10,12 @@ export const falsey_equality = (<T>(v1: T, v2: T) => false) satisfies EqualityFn
 export const parseEquality = <T>(equals: EqualityCheck<T>) => (equals === false ? falsey_equality : (equals ?? default_equality)) satisfies EqualityFn<any>
 
 /** transforms a regular equality check function ({@link SimpleSignalConfig.equals}) into a one that throttles when called too frequently. <br>
- * this means that a singal composed of this as its `equals` function will limit propagating itself further, until at least `delta_time_ms`
+ * this means that a signal composed of this as its `equals` function will limit propagating itself further, until at least `delta_time_ms`
  * amount of time has passed since the last time it was potentially propagated.
  * 
  * @param delta_time_ms the time interval in milliseconds for throttling
  * @param base_equals use an optional customized equality checking function. otherwise the default `prev_value === new_value` comparison will be used
- * @returns a throttled version of the equality checking function which would prematurely return a `true` if called too frequently (ie within `delta_time_ms` interval since the last actual equality cheking)
+ * @returns a throttled version of the equality checking function which would prematurely return a `true` if called too frequently (ie within `delta_time_ms` interval since the last actual equality checking)
  * 
  * @example
  * ```ts