### description:
this pr fixes the previously incorrect topological ordering of the signals' execution. see issue #4 .
in addition, the new implementation is much faster than the previous one (3 times faster in cached-mode, 4 times faster in uncached-mode).

### related issues:
- #4 update cycle is not topologically ordered

### excerpt from git comments:

changes:
- `ids_to_visit_cache_create_new_entry` now generates the list of ids to visit in topological visitation order, while also maintainin that the original source ids are always at the beginning, which is an important piece of info used for a micro-optimization by `fireUpdateCycle`
- each cycle we maintain two dynamic book-keeping sets:
  - `next_to_visit_this_cycle`: which unvisited ids **must** be visited later on (since one of their dependencies have updated)
  - `not_to_visit_this_cycle`: which ids have been declared as aborted by one of their dependencies. (and therefore must never run)
- `fireUpdateCycle`: traverses through the topologically ordered set of ids, executing the signals one-by-one, updating the 2 book-keeping sets, checking of the possiblity of early termination, and then finally running all `postrun`s
- `executeSignal`: simply runs the `run` method of the retrived signal, and stores any `postrun` functions it may contain, then returns the update status of the singnal's `run` method
- add documentation to a bunch of inner functions of the `Context`

improvements:
- the cycle also runs MAD faster now, since we are no longer having to retrieve dependencies of each newly visited ids via `rmap_get` hashmap table.
- for comparison, the benchmark test 2 and 3 in `/test/simple_signal.test.ts` have decreased their average run time to:
  - `"test cached performance"`: from 200ms to 70ms
  - `"test uncached performance"`: from 650ms to 160ms

miscellaneous:
- bump minor version to `0.2.0`
- update dependencies:
  - `kitchensink_ts`: `v0.7.2` -> `v0.7.3`
- fix typos across repo